requirements.txt
amqp==2.4.1
apipkg==1.5
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==18.2.0
Babel==2.6.0
bcrypt==3.1.6
billiard==3.5.0.5
celery==4.2.1
certifi==2018.11.29
cffi==1.12.0
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.2
cryptography==2.5
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.1.7
django-celery-results==1.0.4
django-constance==2.3.1
django-cors-headers==2.4.0
django-environ==0.4.5
django-extensions==2.1.5
django-filter==2.1.0
django-heroku==0.3.1
django-picklefield==2.0
django-redis==4.10.0
django-redis-cache==1.8.1
django-rest-swagger==2.2.0
djangorestframework==3.9.1
execnet==1.5.0
Fabric3==1.14.post1
factory-boy==2.7.0
fake-factory==0.7.4
flower==0.9.2
gunicorn==19.9.0
idna==2.8
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
kombu==4.3.0
MarkupSafe==1.1.0
more-itertools==6.0.0
oauthlib==3.0.1
openapi-codec==1.3.2
packaging==19.0
paramiko==2.4.2
pep8==1.7.1
Pillow==5.4.1
pluggy==0.8.1
psycopg2==2.7.7
py==1.7.0
pyasn1==0.4.5
pycparser==2.19
pyflakes==2.1.0
PyJWT==1.7.1
PyNaCl==1.3.0
pyparsing==2.3.1
pytest==4.2.1
pytest-cache==1.0
pytest-cov==2.6.1
pytest-django==3.4.7
pytest-flakes==4.0.0
pytest-isort==0.2.1
pytest-pep8==1.0.6
pytest-sugar==0.9.2
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2018.9
qrcode==6.1
redis==2.10.6
requests==2.21.0
requests-oauthlib==1.2.0
simplejson==3.16.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.1.0
termcolor==1.1.0
tornado==5.1.1
uritemplate==3.0.0
urllib3==1.24.1
uWSGI==2.0.18
vine==1.2.0
Werkzeug==0.14.1
whitenoise==4.1.2
